addIdCheck

+++ modules/routerFunc.js
@@ -1,6 +1,7 @@
 const Joi = require('joi');
 const contactService = require('../services/contact.service');
 const { createError } = require('./errors');
+const { ObjectId } = require('mongoose').Types;
 
 const schema = Joi.object({
   name: Joi.string().alphanum().min(3).max(30).required(),
@@ -11,7 +12,7 @@ const schema = Joi.object({
     })
     .required(),
   phone: Joi.string().required(),
-  favorite: Joi.boolean(),
+  favorite: Joi.bool(),
 });
 
 const getContacts = async (req, res, next) => {
@@ -26,9 +27,12 @@ const getContacts = async (req, res, next) => {
 const getContactById = async (req, res, next) => {
   try {
     const { contactId } = req.params;
+    if (!ObjectId.isValid(contactId)) {
+      throw createError(404, 'Not found');
+    }
     const contact = await contactService.getContactById(contactId);
     if (!contact) {
-      res.status(404).json({ message: 'Not found' });
+      throw createError(404, 'Not found');
     } else {
       res.json(contact);
     }
@@ -64,6 +68,9 @@ const addContact = async (req, res, next) => {
 const editContact = async (req, res, next) => {
   try {
     const { contactId } = req.params;
+    if (!ObjectId.isValid(contactId)) {
+      throw createError(404, 'Not found');
+    }
     const body = req.body;
     const contact = await contactService.updateContact(contactId, body);
     if (Object.keys(body).length === 0) {
@@ -89,8 +96,10 @@ const editContactFavorite = async (req, res, next) => {
       throw createError(400, 'missing field favorite');
     } else {
       const { contactId } = req.params;
+      if (!ObjectId.isValid(contactId)) {
+        throw createError(404, 'Not found');
+      }
       const contact = await contactService.updateStatusContact(contactId, body);
-      console.log(contact);
       if (!contact) {
         throw createError(404, 'Not found');
       } else {
@@ -104,6 +113,9 @@ const editContactFavorite = async (req, res, next) => {
 const deleteContact = async (req, res, next) => {
   try {
     const { contactId } = req.params;
+    if (!ObjectId.isValid(contactId)) {
+      throw createError(404, 'Not found');
+    }
     const contact = await contactService.removeContact(contactId);
     if (!contact) {
       res.status(404).json({ message: 'Not found' });
